// src/lib/interfaces.ts
export interface VideoMetadata {
  // Here, only assume common aspect ratios - 16:9, 18:9, 9:16 (for youtube shorts, tiktok, instagram reels)
  width: number; // width in pixels
  height: number; // height in pixels
  color: {
    accentColor: string; // hex code
    gradientStartColor?: string; // hex code for the starting color of the gradient
    gradientEndColor?: string; // hex code for the ending color of the gradient
  };
  topic: string; // topic of the video
  description: string; // simple description to put below the youtube video, mentioning the contents of the video
  title: string; // short incentivizing title for youtube video (somewhat clickbaity)
  durationInSeconds: number; // desired duration of the video in seconds
  style: "fun" | "professional" | "normal"; // "professional" is corporate sounding, "fun" is fancy and informal, "normal" is default.

  // ! OMIT GRAPHIC FOR NOW...

  graphic?: {
    // Can there be a relevant graph or chart?
    topic: string; // What should the graph/chart represent?
    type:
      | "line-graph"
      | "histogram"
      | "curve-graph"
      | "bar-graph"
      | "pie-chart";
  };
  table?: {
    // Can a relevant table be displayed?
    label: string; // Short label of what the table represents
  };
}
// ...
export type VideoSection = {
  talkingPoint: string;
  title: string;
  images?: CustomImageDataFromBing[];
  voiceAudio: { urls: [string, string]; durations: [number, number] }; // Max 2 audios allowed.
  music?: { urls: [string]; durations: [number] }; // Max 1 music allowed.
  bulletPoint?: string; // Not decided this yet...
  // variants: string[]; // example ["male-high", "female-sweet", "male heavy", "female-kind" ]
};

export type VideoIntro = {
  // videoHost?: { name: string };
  talkingPoint: string;
  voiceAudio: { urls: [string, string]; durations: [number, number] }; // Max 2 audios allowed.
  music?: { urls: [string]; durations: [number] }; // Max 1 music allowed.
  images?: CustomImageDataFromBing[];
};

export type VideoOutro = {
  // videoHost?: { name: string };
  talkingPoint: string;
  voiceAudio: { urls: [string, string]; durations: [number, number] }; // Max 2 audios allowed.
  music?: { urls: [string]; durations: [number] }; // Max 1 music allowed.
  // images?: CustomImageDataFromBing[];
};

export type VideoTable = {
  table: string;
  isPresent: boolean;
  summary: string;
  voiceAudio: { urls: [string, string]; durations: [number, number] }; // Max 2 audios allowed.
  music?: { urls: [string]; durations: [number] }; // Max 1 music allowed.
};

export type ActualVideoContents = {
  videoSections: VideoSection[];
  intro: VideoIntro;
  outro: VideoOutro;
  table: VideoTable;
  chart?: object[]; // Forget for now... Not planned this yet...
};

export type CustomImageDataFromBing = {
  name: string;
  contentUrl: string;
  hostPageUrl: string;
  hostPageDisplayUrl: string;
  encodingFormat: string;
  contentSize: string;
  width: number;
  height: number;
  accentColor: string;
  imageId: string;
};